+++ nowait-api/src/main/java/com/nowait/application/dto/response/payment/PaymentTokenRes.java
@@ -1,4 +1,4 @@
-package com.nowait.controller.api.dto.response;
+package com.nowait.application.dto.response.payment;
 
 public record PaymentTokenRes(
     String paymentToken
